move styled components
