+++ components/tictac/Board.tsx
@@ -1,20 +1,12 @@
 import { Square } from './Square'
-import styled from 'styled-components'
 import { WProps } from '../../models/Tictac'
-
-const BoardRow = styled.div`
-  &::after {
-    clear: both;
-    content: '';
-    display: table;
-  }
-`
+import { BoardRow } from './Style'
 
 export const Board = (props: { squares: any; onClick: any; winner: WProps | null }) => {
   const win = (i: number): any =>
     props.winner &&
     props.winner.numbers &&
-    props.winner.numbers.map((n: number) => i === n && ' highlight ')
+    props.winner.numbers.map((n: number) => i === n && 'yellow')
 
   const renderSquare = (i: number) => (
     <Square value={props.squares[i]} onClick={() => props.onClick(i)} className={win(i)} />
