+++ components/tictac/Clock.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react'
 import { WProps } from '../../models/Tictac'
+import { H2 } from './Style'
 
 const startTime = Date.now()
-
 const Timer = (currentDate: any, winner: WProps | null) => {
   const [date, setDate] = useState(currentDate)
   const tick = () => setDate(Math.round((Date.now() - startTime) / 1000))
@@ -26,15 +26,7 @@ export const Clock = (props: { winner: WProps | null }) => {
   const timer = Timer(0, props.winner)
   return (
     <div>
-      <h2> {timer} sec has passed </h2>
-      <style>
-        {`
-          h2 {
-            font-size: 20px;
-            margin-top: 15px;
-          }
-        `}
-      </style>
+      <H2> {timer} sec has passed </H2>
     </div>
   )
 }
