+++ styled/Tetris.ts
@@ -1,17 +1,3 @@
-import styled from 'styled-components'
-
-export const BoardWrapperStyle = styled.div`
-  position: relative;
-  width: 300px;
-  background-color: rgb(15, 15, 27);
-  height: 600px;
-`
-export const BoardStyle = styled.div`
-  position: absolute;
-  width: 100%;
-  top: 0px;
-  left: 0px;
-`
 export const btnStyle: React.CSSProperties = {
   margin: '20px 10px',
   width: '100px',
